Fix minor issues with clang-tidy workflow

+++ build-scripts/clang-tidy-run.sh
@@ -7,24 +7,36 @@ set -exo pipefail
 
 num_jobs=3
 
+# enable all the switches by default
+BACKTRACE=${BACKTRACE:-1}
+LOCALIZE=${LOCALIZE:-1}
+TILES=${TILES:-1}
+SOUND=${SOUND:-1}
+
 # create compilation database (compile_commands.json)
-mkdir -p build
-cd build
-cmake \
-    -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
-    ${COMPILER:+-DCMAKE_CXX_COMPILER=$COMPILER} \
-    -DCMAKE_BUILD_TYPE="Release" \
-    -DBACKTRACE=ON \
-    -DTILES=${TILES:-0} \
-    -DSOUND=${SOUND:-0} \
-    -DLOCALIZE=${LOCALIZE:-0} \
-    ..
-cd ..
-ln -s build/compile_commands.json .
+if [ ! -f compile_commands.json ]
+then
+    mkdir -p build
+    cd build
+    cmake \
+        -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
+        ${COMPILER:+-DCMAKE_CXX_COMPILER=$COMPILER} \
+        -DCMAKE_BUILD_TYPE="Release" \
+        -DBACKTRACE=${BACKTRACE} \
+        -DLOCALIZE=${LOCALIZE} \
+        -DTILES=${TILES} \
+        -DSOUND=${SOUND} \
+        ..
+    cd ..
+    ln -s build/compile_commands.json .
+fi
 
 if [ ! -f build/tools/clang-tidy-plugin/libCataAnalyzerPlugin.so ]
 then
     echo "Cata plugin not found. Assuming we're in CI and bailing out."
+    echo "If you are running clang-tidy locally with no plugin, consider"
+    echo "calling it explicitly with the files you care to check."
+    echo 'e.g. `clang-tidy src/item* tests/item*` '
     exit 1
 fi
 
@@ -59,8 +71,10 @@ else
         --silent \
         -j $num_jobs \
         ${COMPILER:+COMPILER=$COMPILER} \
-        TILES=${TILES:-0} \
-        SOUND=${SOUND:-0} \
+        BACKTRACE=${BACKTRACE} \
+        LOCALIZE=${LOCALIZE} \
+        TILES=${TILES} \
+        SOUND=${SOUND} \
         includes
 
     tidyable_cpp_files="$( \
